requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
av==15.0.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.3
ctranslate2==4.4.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
durationpy==0.10
einops==0.8.1
fastapi==0.116.1
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
kubernetes==33.1.0
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.2
lightning-utilities==0.15.0
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
narwhals==1.48.1
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
optuna==4.4.0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
posthog==5.4.0
primePy==1.3
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
semver==3.0.4
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.41
starlette==0.47.2
streamlit==1.47.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
timm==1.0.19
tmdbv3api==1.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.8.0
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.54.0
transnetv2-pytorch==1.0.5
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whisperx @ git+https://github.com/m-bain/whisperx.git@2d9ce44329ae73af2520196d31cd14b6192ace44
yarl==1.20.1
zipp==3.23.0
